+++ packages/layout/CHANGELOG.md
@@ -11,5 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Development dependencies version bumps
 - Direct dependency `antd` has been fixed to `4.24.1` due to a breaking change in its own dependencies (`rc-motion` and `rc-image`) on build
+- Checking [issue](https://github.com/react-component/image/issues/144) for `rc-image` not import `rc-motion`
+- Fixed `monaco-editor` workers both for build and storybook
 
 ## [0.1.7] - 2022-11-15